Use ConfigParser to parse user config file
This fixes a bug caused by extra characters (e.g. newline) leading to an invalid `~/.config/fabric.env`. A newline should be allowed yet causes an exception to be raised with a traceback pointing elsewhere in the code. This could be difficult to fix for some users.

```shell
Traceback (most recent call last):
  File "/Users/ryan/fabric/client/fabric", line 68, in <module>
    standalone.sendMessage(text)
  File "/Users/ryan/fabric/client/utils.py", line 100, in sendMessage
    print(response)
          ^^^^^^^^
UnboundLocalError: cannot access local variable 'response' where it is not associated with a value
```