helper utility for saving a Markdown file
'save' can be used to save a Markdown file, with optional frontmatter
and additional tags. By default, if set, `FABRIC_FRONTMATTER_TAGS` will
be placed into the file as it is written. These tags and front matter
are suppressed from STDOUT, which can be piped into other patterns or
programs with no ill effects. This strives to be a version of `tee` that
is enhanced for personal knowledge systems that use frontmatter.