Adding new pattern: create_recursive_outline.
This pattern is actually based on this incredibly great article: https://learnhowtolearn.org/how-to-build-extremely-quickly/

The idea is to use this pattern whenever you want to break an idea or
task down into small components, fully fleshing out your own TODO list
of things to implement to get it working.

This applies to things like writing articles/papers, creating
applications, and much more.